% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_MLR.R
\name{est.MLR}
\alias{est.MLR}
\title{Estimate age and prediction interval using multiple linear regression}
\usage{
est.MLR(area = 0.95)
}
\arguments{
\item{area}{The desired area for prediction interval
default = 0.95 for 95% prediction interval}
}
\description{
Estimate age and prediction interval using multiple linear regression
}
