% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_MLR.R
\name{est_MLR}
\alias{est_MLR}
\title{Estimate age and prediction interval using multiple linear regression}
\usage{
est_MLR(area = 0.95)
}
\arguments{
\item{area}{The desired area for prediction interval,
default = 0.95 for 95% prediction interval}
}
\description{
This function uses multiple regression to estimate age and
also returns the 95% (default) prediction intervals.
See Brown (1993:Sec. 5.2)
}
\references{
Brown, P. J. (1993). Measurement, Regression, and Calibration.
New York: Oxford University Press.
}
