% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parms.R
\docType{data}
\name{parms}
\alias{parms}
\title{Object containing summary statistics}
\format{
An object of class \code{list} of length 9.
}
\usage{
parms
}
\description{
Object containing summary statistics
}
\keyword{datasets}
