% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_R.R
\name{calc_R}
\alias{calc_R}
\title{Find R (discrepancy statistic) and its probability value}
\usage{
calc_R()
}
\description{
This function finds the R statistic (see Brown 1993: eqn. 5.55)
and its probability value.
}
\references{
Brown, P. J. (1993). Measurement, Regression, and Calibration.
New York: Oxford University Press.
}
