% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Rx.R
\name{calc_Rx}
\alias{calc_Rx}
\title{Find Rx (extrapolation statistic) and its probability value}
\usage{
calc_Rx()
}
\description{
Brown, P. J. (1993). Measurement, Regression, and Calibration.
New York: Oxford University Press.
}
