% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Write x.new that contains the "gpower"
transformed OPD and the average complete
secondary osteon diameter in mm.  This is
interactive and consequenty takes its input
from your enries.}
\usage{
convert()
}
\description{
Write x.new that contains the "gpower"
transformed OPD and the average complete
secondary osteon diameter in mm.  This is
interactive and consequenty takes its input
from your enries.
}
