% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_like.R
\name{profile_like}
\alias{profile_like}
\title{Find profile likelihood estimate and prediction interval for age}
\usage{
profile_like(perc = 95)
}
\arguments{
\item{perc}{Desired prediction interval.  The default value is perc=95,
meaning the 95 percent prediction interval. There is the constraint that
0<perc<100}
}
\value{
returns lower bound of the prediction interval, the estimated
value, and the upper bound of the prediction intervals.
}
\description{
This script finds the profile likelihood estimate of age
and the prediction interval with the desired width.
}
\details{
This script takes as input the gpower transformed OPD and average
complete secondary osteon diameter in mm. from the mid-shaft 6th rib.
This data comes from the vector x.new.  This vector was written by the
interactive script convert().  The profile likelihood method is given in
Brown (1993: Sec. 5.8) and in this case with m=1 (no replication).
}
\references{
Brown, P. J. (1993). \emph{Measurement, Regression, and Calibration}.
New York: Oxford University Press.
82(2), 199-220.
}
